Test runtime profile pytest plugin behavior
