+++ tests/test_runtime_profiles.py
@@ -1,11 +1,15 @@
 from __future__ import annotations
 
+import pytest
+
 from worldforge.testing.runtime_profiles import (
     PROVIDER_RUNTIME_PROFILES_BY_NAME,
     provider_profile_skip_reason,
     pytest_marker_definitions,
     runtime_marker_skip_reason,
 )
+
+pytest_plugins = ("pytester",)
 
 
 def test_runtime_markers_require_explicit_opt_in() -> None:
@@ -48,3 +52,64 @@ def test_marker_definitions_cover_runtime_profiles() -> None:
         "gr00t",
         "lerobot",
     }
+
+
+def test_pytest_plugin_skips_live_tests_by_default(pytester: pytest.Pytester) -> None:
+    pytester.makeconftest('pytest_plugins = ("worldforge.testing.pytest_plugin",)')
+    pytester.makepyfile("""
+        import pytest
+
+        @pytest.mark.live
+        def test_live_runtime():
+            raise AssertionError("should be skipped before test body runs")
+    """)
+
+    result = pytester.runpytest("-rs")
+
+    result.assert_outcomes(skipped=1)
+    result.stdout.fnmatch_lines(["*requires an explicit --run-live opt-in*"])
+
+
+def test_pytest_plugin_runs_configured_provider_profile(
+    pytester: pytest.Pytester,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    pytester.makeconftest('pytest_plugins = ("worldforge.testing.pytest_plugin",)')
+    pytester.makepyfile("""
+        import pytest
+
+        @pytest.mark.live
+        @pytest.mark.network
+        @pytest.mark.credentialed
+        @pytest.mark.provider_profile("runway")
+        def test_runway_runtime():
+            assert True
+    """)
+    monkeypatch.setenv("RUNWAY_API_SECRET", "test-secret")
+
+    result = pytester.runpytest(
+        "--run-live",
+        "--run-network",
+        "--run-credentialed",
+        "--provider-profile",
+        "runway",
+    )
+
+    result.assert_outcomes(passed=1)
+
+
+def test_pytest_plugin_skips_unselected_provider_profile(pytester: pytest.Pytester) -> None:
+    pytester.makeconftest('pytest_plugins = ("worldforge.testing.pytest_plugin",)')
+    pytester.makepyfile("""
+        import pytest
+
+        @pytest.mark.live
+        @pytest.mark.provider_profile("runway")
+        def test_runway_runtime():
+            raise AssertionError("should be skipped before test body runs")
+    """)
+
+    result = pytester.runpytest("-rs", "--run-live", "--provider-profile", "cosmos")
+
+    result.assert_outcomes(skipped=1)
+    result.stdout.fnmatch_lines(["*selected provider profile is 'cosmos', not 'runway'*"])
